Compute logprobs in fp32 for HF and MLX backends
`torch.log_softmax(bf16_in)` returns a bf16 tensor; the per-element bf16
storage precision loses information in the tail of the distribution.
Measured impact on Qwen2.5-1.5B-Instruct (bf16) logprobs, vs the same
backend with fp32 softmax:

  bin (logprob range)   HF L∞   MLX L∞
  head  [-2.3, 0]       0.008   0.062
  upper [-6.9, -2.3)    0.020   0.062
  mid   [-13.8, -6.9)   0.037   0.062
  tail  [-27.6, -13.8)  0.069   0.124
  deep  [-69.1, -27.6)  0.126   0.187

vLLM and SGL already do the softmax in fp32 internally; no change there.

+++ genlm/backend/llm/mlx.py
@@ -342,7 +342,7 @@ else:
                 inputs_padded = inputs_padded[:, n_to_process:]
 
             logits = self.mlx_lm_model(inputs_padded, cache=prompt_cache)
-            logits = logits[:, -1, :]
+            logits = logits[:, -1, :].astype(mx.float32)
             logprobs = logits - mx.logsumexp(logits, axis=-1, keepdims=True)
             mx.async_eval(logprobs)
